simpler jsonifier
successfully jsonified every algorithm under algs-c and algs-py

produces "levels.json" when "jsonifier.js" is run, containing 3 layers:
category --> language --> (lines, comments)

"category" is bubblesort, mergesort, etc.

"language" is either "c" or "py" as of right now.

"lines" is within language and is the "category.language"'s code lines.
for example, this could be bubblesort.c's lines of code. every line is
placed into a JSON with its line number being how you access it. as of
right now, the line number is the literal line number on the algorithm
itself.

"comments" is also within language and is the comments of
"category.language". every comment is placed into a JSON and assigned
the line number that it's a comment for.